--- showfolio/src/components/Categories.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Categories = () => <div>
-
-</div>
-
-export default Categories;
