--- showfolio/src/styles/Project.module.scss
@@ -1,13 +1,0 @@
-@import './Colors.scss';
-
-
-.project {
-  height: 284px;
-  width: 1020px;
-}
-
-.body {
-  background-color: $white;
-  height: 250px;
-  border-radius: 10px;
-}
